yatri/requirements.txt
brotlipy==0.7.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.3
django-allauth==0.35.0
django-cors-headers==2.2.0
django-paypal==0.5.0
django-pipeline==1.6.14
djangorestframework==3.7.7
idna==2.6
oauthlib==2.0.7
pycparser==2.18
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
stripe==1.79.1
urllib3==1.22
whitenoise==3.3.1
